Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int a;
-    scanf("%ld",&a);
+    float a;
+    scanf("%f",&a);
     if (a & -1)
     {
         printf("Set");
